git-svn-id: https://chibios.svn.sourceforge.net/svnroot/chibios/branches/stm8_cosmic_dev@1981 35acf78f-673a-0410-8e92-d51de3d6d3f4

+++ os/ports/cosmic/STM8/chtypes.h
@@ -94,7 +94,7 @@ typedef uint16_t        systime_t;
 /**
  * @brief   Counter, recommended fastest signed.
  */
-typedef int16_t         cnt_t;
+typedef int8_t          cnt_t;
 
 /**
  * @brief   Inline function modifier.
